[ZDF/3sat/phoenix] Let ZDF extractor work with 3sat.de
This patch also fixes the brokenness of 3sat.
